Rename channel metric collector
