+++ ibm_mq/datadog_checks/ibm_mq/ibm_mq.py
@@ -8,7 +8,7 @@ from datadog_checks.base import AgentCheck
 from datadog_checks.ibm_mq.metrics import COUNT, GAUGE
 
 from . import connection, errors
-from .channel_metric_collection import ChannelMetricCollector
+from .channel_metric_collector import ChannelMetricCollector
 from .config import IBMMQConfig
 from .queue_metric_collector import QueueMetricCollector
 
@@ -34,7 +34,7 @@ class IbmMqCheck(AgentCheck):
         self.queue_metric_collector = QueueMetricCollector(
             self.config, self.service_check, self.warning, self.send_metric, self.log
         )
-        self.channel_metric_collection = ChannelMetricCollector(self.config, self.service_check, self.gauge, self.log)
+        self.channel_metric_collector = ChannelMetricCollector(self.config, self.service_check, self.gauge, self.log)
 
     def check(self, _):
         try:
@@ -46,7 +46,7 @@ class IbmMqCheck(AgentCheck):
             return
 
         try:
-            self.channel_metric_collection.get_pcf_channel_metrics(queue_manager)
+            self.channel_metric_collector.get_pcf_channel_metrics(queue_manager)
             self.queue_metric_collector.collect_queue_metrics(queue_manager)
         finally:
             queue_manager.disconnect()
