+++ ibm_mq/tests/test_ibm_mq_unit.py
@@ -32,7 +32,7 @@ def test_channel_status_service_check_default_mapping(aggregator, instance):
     }
 
     for status in service_check_map:
-        check.channel_metric_collection._submit_status_check(
+        check.channel_metric_collector._submit_status_check(
             'my_channel', status, ["channel:my_channel_{}".format(status)]
         )
 
@@ -75,7 +75,7 @@ def test_channel_status_service_check_custom_mapping(aggregator, instance):
     }
 
     for status in service_check_map:
-        check.channel_metric_collection._submit_status_check(
+        check.channel_metric_collector._submit_status_check(
             'my_channel', status, ["channel:my_channel_{}".format(status)]
         )
